LF-4164 Move details subcomponent cards into their own folder
