+++ packages/webapp/src/components/Animals/DetailCards/General.tsx
