+++ packages/webapp/src/components/Animals/DetailCards/styles.module.scss
@@ -13,26 +13,23 @@
 *  GNU General Public License for more details, see <https://www.gnu.org/licenses/>.
 */
 
-.section {
-  padding: 16px;
-  background: var(--White);
-  border-radius: 4px;
-  border: 1px solid var(--Colors-Neutral-Neutral-100);
-}
-
-.expanded {
-  border: 1px solid var(--Form-focus);
-  box-shadow: 0px 0px 4px 0px var(--Colors-Neutral-Neutral-200);
-}
-
-.expandedContentWrapper {
-  margin-top: 16px;
-}
-
-.detailsWrapper {
+.sectionWrapper {
   display: flex;
   flex-direction: column;
   gap: 8px;
   padding: 16px;
   background: var(--White);
+}
+
+.countAndSexDetailsWrapper {
+  display: flex;
+  align-items: start;
+  gap: 16px;
+}
+
+.countInput {
+  flex-basis: 104px;
+  & + div {
+    flex: 1;
+  }
 }
